MAILBOX-262 try to make Abstract*Message useless by delegating when needed
