+++ mailbox/jpa/src/main/java/org/apache/james/mailbox/jpa/mail/model/openjpa/JPAEncryptedMailboxMessage.java
@@ -61,9 +61,6 @@ public class JPAEncryptedMailboxMessage extends AbstractJPAMailboxMessage {
         @Factory("EncryptDecryptHelper.getDecrypted")
         @Lob private byte[] header;
         
-        @Deprecated
-        public JPAEncryptedMailboxMessage() {}
-
         public JPAEncryptedMailboxMessage(JPAMailbox mailbox, Date internalDate, int size, Flags flags, SharedInputStream content, int bodyStartOctet, final PropertyBuilder propertyBuilder) throws MailboxException {
             super(mailbox, internalDate, flags, size ,bodyStartOctet, propertyBuilder);
             try {
@@ -81,9 +78,6 @@ public class JPAEncryptedMailboxMessage extends AbstractJPAMailboxMessage {
 
         /**
          * Create a copy of the given message
-         * 
-         * @param message
-         * @throws MailboxException 
          */
         public JPAEncryptedMailboxMessage(JPAMailbox mailbox, long uid, long modSeq, MailboxMessage<?> message) throws MailboxException{
             super(mailbox, uid, modSeq, message);
@@ -96,16 +90,12 @@ public class JPAEncryptedMailboxMessage extends AbstractJPAMailboxMessage {
         }
 
 
-        /**
-         * @see MailboxMessage#getBodyContent()
-         */
+        @Override
         public InputStream getBodyContent() throws IOException {
             return new ByteArrayInputStream(body);
         }
 
-        /**
-         * @see MailboxMessage#getHeaderContent()
-         */
+        @Override
         public InputStream getHeaderContent() throws IOException {
             return new ByteArrayInputStream(header);
         }
