+++ mailbox/jpa/src/main/java/org/apache/james/mailbox/jpa/mail/model/openjpa/JPAStreamingMailboxMessage.java
@@ -61,9 +61,6 @@ public class JPAStreamingMailboxMessage extends AbstractJPAMailboxMessage {
 
     private SharedInputStream content;
 
-    @Deprecated
-    public JPAStreamingMailboxMessage() {}
-
     public JPAStreamingMailboxMessage(JPAMailbox mailbox, Date internalDate, int size, Flags flags, SharedInputStream content, int bodyStartOctet, final PropertyBuilder propertyBuilder) throws MailboxException {
         super(mailbox, internalDate, flags, size ,bodyStartOctet, propertyBuilder);
         this.content = content;
@@ -79,9 +76,6 @@ public class JPAStreamingMailboxMessage extends AbstractJPAMailboxMessage {
 
     /**
      * Create a copy of the given message
-     * 
-     * @param message
-     * @throws IOException 
      */
     public JPAStreamingMailboxMessage(JPAMailbox mailbox, long uid, long modSeq, MailboxMessage<?> message) throws MailboxException {
         super(mailbox, uid, modSeq, message);
@@ -94,17 +88,12 @@ public class JPAStreamingMailboxMessage extends AbstractJPAMailboxMessage {
         }
     }
 
-    
-    /**
-     * @see MailboxMessage#getBodyContent()
-     */
+    @Override
     public InputStream getBodyContent() throws IOException {
         return content.newStream(getBodyStartOctet(), -1);
     }
 
-    /**
-     * @see MailboxMessage#getHeaderContent()
-     */
+    @Override
     public InputStream getHeaderContent() throws IOException {
         int headerEnd = getBodyStartOctet() -2;
         if (headerEnd < 0) {
